% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exonsToIntrons.R
\name{exonsToIntrons}
\alias{exonsToIntrons}
\title{Convert exon annotation data.frame to intron locations}
\usage{
exonsToIntrons(exons, max_dist = 50)
}
\arguments{
\item{exons}{data.frame containing exon co-ordinates.
Should be produced by gtfToExons()}

\item{max_dist}{Maximum distance from the 3' exon to create the branchpoint region.}
}
\value{
vector of chromosome names and intronic locations
}
\description{
Converts exon annotation to intron locations overlapping the branchpoint region
for exculsion of non-branchpoint region SNPs
Returns a character vector of chromosome_location
}
\examples{
exons <- readExonAnnotation("gencode.v24.exons.txt")
introns <- exonsToIntrons(exons, 50)
}
\author{
Beth Signal
}

