% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExonDists.R
\name{getExonDists}
\alias{getExonDists}
\title{Get the closest 3' and 5' exons}
\usage{
getExonDists(query_line, exon_annotation, query_type)
}
\arguments{
\item{query_line}{line from a query data.frame (or a character vector)
must have chromosome at position 2, genomic co-ordinate at position 3,
and starnd at position 4.}

\item{exon_annotation}{data.frame containing exon co-ordinates.
Should be produced by gtfToExons()}

\item{query_type}{type of query. "SNP" or "region"}
}
\value{
vector with distance to the closest 3' and 5' exons,
whether these exons are part of the same gene (i.e is the location intronic),
and the identifiers for the 3' and 5' exons.
}
\description{
Finds the closest annotated exons from a genomic co-ordinate.
Returns the distance to the 3' exon, distance to the 5' exon,
ids of the 3' and 5' exon,
and if the exons are from the same parent gene
}
\examples{
exons_dists <- getExonDists(query_line, exons)
}
\author{
Beth Signal
}

