% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGtfAttribute.R
\name{getGtfAttribute}
\alias{getGtfAttribute}
\title{extract a field from the GTF annotation}
\usage{
getGtfAttribute(gtf, attribute)
}
\arguments{
\item{gtf}{gtf annotation data.frame}

\item{attribute}{name of the attribute to exact the value for}
}
\value{
vector with attribute value, or NA if attribute name is not found.
}
\description{
extract a field from the GTF annotation (usually column 9)
}
\examples{
gtf <- as.data.frame(data.table::fread(system.file("extdata","gencode.v24.annotation.small.gtf", package = "branchpointer")))
gtf$gene_id <- getGtfAttribute(gtf, "gene_id")
}
\author{
Beth Signal
}

