% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPPT.R
\name{getPPT}
\alias{getPPT}
\title{Get the best polypyramidine tract}
\usage{
getPPT(attributes_line)
}
\arguments{
\item{attributes_line}{line from a query attributes data.frame (or a character vector)
must have 3'SS distance at 3,
and nucleotide sequence at position 4.}
}
\value{
distance to the start of the longest PPT, and its length
}
\description{
Takes a query genomic sequence, finds all potential polypyramidine tracts (PPTs)
between the test site and the annotated 3'exon.
Returns the distance to the start of the longest PPT, and its length.
}
\examples{
attributes_line <- c("id", "1000","27", paste(c("A","G",rep(c("C","T"),4))[sample(1:10, 501, replace=TRUE)] , collapse=""))
pyra_df <- getPPT(attributes_line)
}
\author{
Beth Signal
}

