% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQueryLoc.R
\name{getQueryLoc}
\alias{getQueryLoc}
\title{Find the closest 3' and 5' exons to a branchpointer query}
\usage{
getQueryLoc(query, query_type, max_dist = 50, filter = TRUE, exons,
  use_parallel = FALSE, cores = 1)
}
\arguments{
\item{query}{branchpointer query data.frame
must have chromosome at position 2, genomic co-ordinate at position 3,
and strand at position 4.}

\item{query_type}{type of query file (\code{"SNP"} or \code{"region"})}

\item{max_dist}{maximum distance a SNP can be from an annotated 3' exon.}

\item{filter}{remove SNP queries prior to finding finding nearest exons.}

\item{exons}{data.frame containing exon co-ordinates.
Should be produced by format_exon_annotations.sh}

\item{use_parallel}{use parallelisation to speed up code?
(reccomended for > 500 query entries)}

\item{cores}{number of cores to use in parallelisation (default = \code{1})}
}
\value{
data.frame with the query and its location relative to the 3' and 5' exons
}
\description{
Finds the closest annotated exons from genomic co-ordinates in a branchpointer query data.frame.
}
\examples{
exons <- readExonAnnotation("gencode.v24.exons.txt")
query <- readQueryFile("SNP_example.txt",query_type = "SNP")
query <- getQueryLoc(query,query_type="SNP",exons = exons, filter=FALSE)
}
\author{
Beth Signal
}

