% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBranchpointWindow.R
\name{plotBranchpointWindow}
\alias{plotBranchpointWindow}
\title{Plots branchpointer predictions}
\usage{
plotBranchpointWindow(query_name, branchpoint_predictions, query_attributes,
  probability_cutoff = 0.5, plot_mutated = FALSE, plot_structure = TRUE,
  exons)
}
\arguments{
\item{query_name}{query id used to identify the SNP or region}

\item{branchpoint_predictions}{prediction of branchpoint probabilities for each site}

\item{query_attributes}{attribute data.frame}

\item{probability_cutoff}{probability score cutoff value for displaying U2 binding energy}

\item{plot_mutated}{plot alternative sequence predicitons alongside reference sequence predictions}

\item{plot_structure}{plot structures for gene and 3' exon containing and skipping isoforms}

\item{exons}{data.frame containing exon co-ordinates.
Should be produced by format_exon_annotations.sh}
}
\value{
ggplot2 plot with branchpoint features in the specified intronic region
}
\description{
Plots branchpointer predictions
}
\examples{
exons <- readExonAnnotation("gencode.v24.exons.txt")
query <- readQueryFile("SNP_example.txt",query_type = "SNP")
query <- getQueryLoc(query,query_type="SNP",exons = exons, filter=FALSE)
query_attributes <- getBranchpointSequence(query,query_type = "SNP",
genome = "GRCh38.p5.genome.fa",
bedtools_location="/Applications/bedtools2/bin/bedtools")
branchpoint_predictions <- predictBranchpoints(query_attributes)
plotBranchpointWindow(query$id[1],  branchpoint_predictions,
query_attributes, plot_mutated = TRUE, exons = exons)
}
\author{
Beth Signal
}

