% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictBranchpoints.R
\name{predictBranchpoints}
\alias{predictBranchpoints}
\title{Predict branchpoint probability scores}
\usage{
predictBranchpoints(query_attributes, use_parallel = FALSE, cores = 1)
}
\arguments{
\item{query_attributes}{data.frame produced by getBranchpointSequence
containing features for branchpoint prediction}

\item{use_parallel}{use parallelisation to speed up code?
(reccomended for > 500 query entries)}

\item{cores}{number of cores to use in parallelisation (default = \code{1})}
}
\value{
data.frame with branchpoint probaility scores for each site in query_attributes
}
\description{
predicts branchpoint probability scores for each query site.
}
\examples{
branchpoint_predicitons <- predictBranchpoints(query_attributes)
}
\author{
Beth Signal
}

