% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionsToStats.R
\name{predictionsToStats}
\alias{predictionsToStats}
\title{Convert SNP branchpoint predictions across the branchpoint window to an intronic summary}
\usage{
predictionsToStats(predictions, query, BP_probability_cutoff = 0.5,
  BP_probability_change = 0.2)
}
\arguments{
\item{predictions}{site-wide branchpoint proability predictions
produced from predictBranchpoints()}

\item{query}{query data.frame containing all snp ids to be summarised}

\item{BP_probability_cutoff}{Value to be used as the cutoff for
discriminating branchpoint sites from non-branchpoint sites
(default = \code{0.5})}

\item{BP_probability_change}{Minimum probability score change
required to call a branchpoint site as deleted or created by
a SNP (default = \code{0.2})}
}
\value{
data.frame with summarised branchpoint changes
occuring within the intron due to a SNP.
}
\description{
Takes predictions of branchpoint probabilities from a reference
and alternative SNP and summarises the effect(s) of the SNP.
}
\examples{
small_exons <- system.file("extdata","gencode.v24.annotation.exons.small.txt",
package = "branchpointer")
exons <- readExonAnnotation(small_exons)
genome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38

query_snp <- system.file("extdata","SNP_example.txt", package = "branchpointer")
query <- readQueryFile(query_snp,query_type = "SNP")
query <- getQueryLoc(query,query_type = "SNP",exons = exons, filter = FALSE)
query_attributes <- getBranchpointSequence(query,
query_type = "SNP",
useBSgenome = TRUE,
BSgenome = genome)
branchpoint_predictions <- predictBranchpoints(query_attributes)
snp_stats <- predictionsToStats(branchpoint_predictions, query)
}
\author{
Beth Signal
}

