% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfToExons.R
\name{gtfToExons}
\alias{gtfToExons}
\title{Convert GTF file to exon location file}
\usage{
gtfToExons(gtf, output_file = NA)
}
\arguments{
\item{gtf}{file containing the gtf annotation.}

\item{output_file}{name of the file to write the exon annotation to.
If not specified, will write to the same location as the gtf file.}
}
\value{
exon annotation data.frame
}
\description{
Converts a GTF annotation to exon locations
}
\examples{
exons <- gtfToExons("~/Downloads/gencode.v24.annotation.small.gtf", output_file="gencode.v24.exons.txt")
}
\author{
Beth Signal
}

