% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readExonAnnotation.R
\name{readExonAnnotation}
\alias{readExonAnnotation}
\title{Read in exon annotation}
\usage{
readExonAnnotation(exon_file, output_file = NA)
}
\arguments{
\item{exon_file}{exon annotation file generated by gtfToExons}

\item{output_file}{name of the file to write the exon annotation to when a .gtf file is specified.
If not specified, will write to the same location as the gtf file.}
}
\value{
exon annotation data.frame
}
\description{
Reads in an exon annotation file generated by gtfToExons
Alternatively, will generate then read the exon annotation file if a .gtf file is specified
}
\examples{
exons <- readExonAnnotation("~/Downloads/gencode.v24.annotation.small.gtf")
small_exons <- system.file("extdata","gencode.v24.annotation.exons.small.txt",
package = "branchpointer")

exons <- readExonAnnotation(small_exons)
}
\author{
Beth Signal
}

